refactored project structure and renamed some test classes
