+++ src/main/java/genius/scores/ScoresTableModel.java
@@ -1,4 +1,4 @@
-package genius.model;
+package genius.scores;
 
 import javax.swing.table.AbstractTableModel;
