+++ src/main/java/genius/settings/Difficulty.java
@@ -6,7 +6,7 @@
  * Created : May 13, 2018
  * Author  : 
  */
-package genius.model;
+package genius.settings;
 
 
 /**
