+++ src/test/java/genius/game/GameTest.java
@@ -6,18 +6,20 @@
  * Created : 14 de mai de 2018
  * Author  : Graeff
  */
-package genius.model;
+package genius.game;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
+
+import genius.button.Button;
 
 
 /**
  * @author Graeff
  *
  */
-public class TestGame {
+public class GameTest {
 	@Test
 	public void enumSameAsInt() {
 		assertEquals(Button.RED, Button.values()[0]);
